Admin page merged with clubpage, bug fixes
