+++ src/pages/ClubPage.js
@@ -12,25 +12,26 @@ import ReactGA from 'react-ga';
 import RightArrow from '@material-ui/icons/CallMadeRounded';
 import HeartBordered from '@material-ui/icons/FavoriteBorderRounded';
 import EditIcon from '@material-ui/icons/EditRounded';
+import { Route, Switch, Link } from 'react-router-dom';
 
 function ClubPage({
+  admin,
   organization,
   getOrganization,
   clearOrganization,
   tagOptions,
   history,
 }) {
-  const routeId = history.location.pathname.slice(6);
+  const path = history.location.pathname.split("/").slice(2);
+  const routeId = path[0];
   useEffect(() => {
-    if (organization.link_name !== routeId) getOrganization(routeId);
+    if (!admin && organization.link_name !== routeId) getOrganization(routeId);
     // clears the loaded profile when component unmounts
     return () => {
       !organization.link_name && clearOrganization();
     };
     // recall useEffect when the link_name in url changes
   }, [routeId]);
-
-  let admin = true;
   /*
   organization.gallery = [
     {
@@ -50,6 +51,10 @@ function ClubPage({
     },
   ];*/
   let [tab, setTab] = useState('overview')
+  const tempTab = path[1];
+  if (tempTab) {
+    tab = tempTab;
+  } 
 
   if (!organization.link_name) return <Loading />;
 
@@ -85,24 +90,49 @@ function ClubPage({
     </div>
   ));
 
+  const overview = 
+    <div>
+    {organization.about_us &&
+      <div className='clubpage-content-about clubpage-content-item' >
+        <div className='clubpage-content-header'>
+          <h1>About {organization.name}</h1>
+          {admin && 
+            <EditIcon className="clubpage-content-header-icon"/>
+          }
+        </div>
+        <p dangerouslySetInnerHTML={{ __html: organization.about_us }}></p>
+      </div>
+    }
+    {organization.gallery &&
+      <div className='clubpage-content-gallery clubpage-content-item' >
+        <div className='clubpage-content-header'>
+          <h1>Gallery</h1>
+          {admin && 
+            <EditIcon className="clubpage-content-header-icon"/>
+          }
+        </div>
+        <Gallery data={organization.gallery}/>
+      </div>
+    }
+  </div>
 
   let categoryList = organization.tags.map((tag, i) => (
-    <Tag label={tagOptions[tag] && tagOptions[tag].label} />
+    <Tag key={i} label={tagOptions[tag] && tagOptions[tag].label} />
   ));
   let tagList = [];
   if (organization.new_members) {
-    tagList.push(<Tag label="Taking New Members" color="#c9f0c9" />);
+    tagList.push(<Tag key={"nm"} label="Taking New Members" color="#c9f0c9" />);
   } else {
-    tagList.push(<Tag label="Not Taking New Members" color="#ffd6d6" />);
+    tagList.push(<Tag key={"nnm"} label="Not Taking New Members" color="#ffd6d6" />);
   }
   if (organization.app_required) {
-    tagList.push(<Tag label="Application Required" color="#fff1ae" />);
+    tagList.push(<Tag key={"ar"} label="Application Required" color="#fff1ae" />);
   } else {
-    tagList.push(<Tag label="Application Not Required" color="#cdeaff" />)
+    tagList.push(<Tag key={"nar"} label="Application Not Required" color="#cdeaff" />)
   }
 
   ReactGA.initialize('UA-176775736-1');
-  ReactGA.pageview('/' + history.location.pathname.slice(6));
+  ReactGA.pageview('/' + history.location.pathname.slice(6).split("/")[0]);
 
   return (
     <div className='clubpage-wrapper'>
@@ -124,11 +154,11 @@ function ClubPage({
               />
             </div>
             <div className="clubpage-header-middle">
-              <div class="club-title">{organization.name}</div>
-              <div class="header-tags">
+              <div className="club-title">{organization.name}</div>
+              <div className="header-tags">
                 {categoryList}
               </div>
-              <div class="header-tags">
+              <div className="header-tags">
                 {tagList}
               </div>
             </div>
@@ -140,66 +170,66 @@ function ClubPage({
             </div>
           </div>
           <div className="clubpage-header-nav">
-            <button className={`clubpage-header-nav-item ${tab == "overview" ? "selected" : ""}`} onClick={() => setTab("overview")}>Overview</button>
-            <button className={`clubpage-header-nav-item ${tab == "recruitment" ? "selected" : ""}`} onClick={() => setTab("recruitment")}>Recruitment</button>
-            <button className={`clubpage-header-nav-item ${tab == "events" ? "selected" : ""}`} onClick={() => setTab("events")}>Events</button>
+            <Link 
+              to={admin ? "/admin/overview" : `/club/${routeId}/overview`}
+              className={`clubpage-header-nav-item ${tab === "overview" ? "selected" : ""}`} onClick={() => setTab("overview")}>
+              Overview
+            </Link>
+            <Link 
+              to={admin ? "/admin/recruitment" : `/club/${routeId}/recruitment`}
+              className={`clubpage-header-nav-item ${tab === "recruitment" ? "selected" : ""}`} onClick={() => setTab("recruitment")}>
+              Recruitment
+            </Link>
+            <Link 
+              to={admin ? "/admin/events" : `/club/${routeId}/events`}
+              className={`clubpage-header-nav-item ${tab === "events" ? "selected" : ""}`} onClick={() => setTab("events")}>
+              Events
+            </Link>
           </div>
         </div>
         <div className='clubpage-content'>
           <div className='clubpage-content-left'>
-            {tab === 'overview' &&
-              <div>
-                {organization.about_us &&
-                  <div className='clubpage-content-about clubpage-content-item' >
-                    <a className='clubpage-content-header'>
-                      <h1>About {organization.name}</h1>
+            <Switch>
+              <Route path={admin ? "/admin/overview" : `/club/${routeId}/overview`} render={() => overview}/>
+              <Route path={admin ? "/admin/recruitment" : `/club/${routeId}/recruitment`} render={() => 
+                <div className= "clubpage-content-timeline">
+                  <div className='clubpage-content-header'>
+                    <h1>Recruitment Timeline</h1>
+                    {admin && 
                       <EditIcon className="clubpage-content-header-icon"/>
-                    </a>
-                    <p dangerouslySetInnerHTML={{ __html: organization.about_us }}></p>
+                    }
                   </div>
-                }
-                {organization.gallery &&
-                  <div className='clubpage-content-gallery clubpage-content-item' >
-                    <a className='clubpage-content-header'>
-                      <h1>Gallery</h1>
-                      <EditIcon className="clubpage-content-header-icon"/>
-                    </a>
-                    <Gallery data={organization.gallery}/>
-                  </div>
-                }
-              </div>
-            }
-            {tab === 'recruitment' && 
-              <div className= "clubpage-content-timeline">
-                <a className='clubpage-content-header'>
-                  <h1>Recruitment Timeline</h1>
-                  <EditIcon className="clubpage-content-header-icon"/>
-                </a>
-              </div>
-            }
-            {tab === 'events' &&
-              <div className= "clubpage-content-events">
-                {organization.events.length > 0 ?
+                </div>
+              } />
+              <Route path={admin ? "/admin/events" : `/club/${routeId}/events`} render={() =>
+                <div className= "clubpage-content-events">
                   <div>
-                    <a className='clubpage-content-header'>
+                    <div className='clubpage-content-header'>
                       <h1>Events</h1>
-                      <EditIcon className="clubpage-content-header-icon"/>
-                    </a>
-                    <EventAccord data={organization} />
+                      {admin && 
+                        <EditIcon className="clubpage-content-header-icon"/>
+                      }
+                      </div>
+                      {organization.events.length > 0 ? 
+                        <EventAccord data={organization} />
+                      :
+                      <p>There are no events scheduled.</p>
+                      }
                   </div>
-                : 
-                  <p>There are no events scheduled.</p>
-                }
-              </div>
-            }
+                </div>
+              } />
+              <Route path={admin ? "/admin" : `/club/${routeId}`} render={() => overview}/>
+            </Switch>
           </div>
           <div className='clubpage-content-right'>
             {organization.get_involved && 
               <div className="clubpage-content-getinvolved clubpage-tile">
-                <a className='clubpage-content-header'>
+                <div className='clubpage-content-header'>
                   <h1>How to Get Involved</h1>
-                  <EditIcon className="clubpage-content-header-icon"/>
-                </a>
+                  {admin && 
+                    <EditIcon className="clubpage-content-header-icon"/>
+                  }
+                  </div>
                 <p>{organization.get_involved}</p>
                 <button className="clubpage-apply-btn" /* NEED AN ONCLICK HANDLER TO LINK TO APPLICATION*/>
                   Apply Now!
@@ -208,10 +238,12 @@ function ClubPage({
               </div>
             }
             <div className="clubpage-content-contact clubpage-tile">
-              <a className='clubpage-content-header'>
+              <div className='clubpage-content-header'>
                 <h1>Contact Information</h1>
-                <EditIcon className="clubpage-content-header-icon"/>
-              </a>
+                {admin && 
+                  <EditIcon className="clubpage-content-header-icon"/>
+                }
+                </div>
               <h2>Website</h2>
               <h2>Email</h2>
               <h2>Social Media</h2>
@@ -219,10 +251,12 @@ function ClubPage({
             </div>
             {organization.resources && organization.resources.length > 0 && 
               <div className="clubpage-content-getinvolved clubpage-tile">
-                <a className='clubpage-content-header'>
+                <div className='clubpage-content-header'>
                   <h1>Resources</h1>
-                  <EditIcon className="clubpage-content-header-icon"/>
-                </a>
+                  {admin && 
+                    <EditIcon className="clubpage-content-header-icon"/>
+                  }
+                </div>
                 <div className="clubpage-content-resource-list">{resComps}</div>
               </div>
             }
@@ -234,11 +268,10 @@ function ClubPage({
   );
 }
 
-const mapStateToProps = (state) => ({
-  organization: state.catalog.organization,
+const mapStateToProps = (state, ownProps) => ({
+  admin: ownProps.admin,
+  organization: ownProps.admin ? state.profile.profile : state.catalog.organization,
   tagOptions: state.profile.tagOptions,
 });
 
-export default connect(mapStateToProps, { getOrganization, clearOrganization })(
-  withRouter(ClubPage)
-);
+export default connect(mapStateToProps, { getOrganization, clearOrganization })(withRouter(ClubPage));
