+++ src/pages/admin/Admin.js
@@ -1,102 +1,26 @@
 import React, { useEffect } from 'react';
 import './Admin.css';
 import { loadProfile } from '../../actions/profile';
-import { Route, Switch, Link } from 'react-router-dom';
 import { connect } from 'react-redux';
 import Profile from './Profile';
 import ContactInfo from './ContactInfo';
 import GetInvolved from './GetInvolved';
 import Events from './Events';
 import Resources from './Resources';
+import ClubPage from '../ClubPage';
 
-const Admin = ({ profile, events, resources, loadProfile }) => {
+const Admin = ({ profile, admin, loadProfile }) => {
   useEffect(() => {
     if (profile && profile.link_name && profile.link_name.length === 0) loadProfile();
   }, [loadProfile, profile]);
 
   return (
-    <div className="clubEdit">
-      <div className="admin-page">
-        <div className="admin-sidebar">
-          <Link
-            to="/admin/profile"
-            className={
-              window.location.pathname === '/admin/profile' ||
-              window.location.pathname === '/admin'
-                ? 'selected page-link'
-                : 'page-link'
-            }
-          >
-            Profile
-          </Link>
-          <Link
-            to="/admin/contact"
-            className={
-              window.location.pathname === '/admin/contact'
-                ? 'selected page-link'
-                : 'page-link'
-            }
-          >
-            Contact Information
-          </Link>
-          <Link
-            to="/admin/getinvolved"
-            className={
-              window.location.pathname === '/admin/getinvolved'
-                ? 'selected page-link'
-                : 'page-link'
-            }
-          >
-            How to Get Involved
-          </Link>
-          <Link
-            to="/admin/resources"
-            className={
-              window.location.pathname === '/admin/resources'
-                ? 'selected page-link'
-                : 'page-link'
-            }
-          >
-            Resources
-          </Link>
-          <Link
-            to="/admin/events"
-            className={
-              window.location.pathname === '/admin/events'
-                ? 'selected page-link'
-                : 'page-link'
-            }
-          >
-            Events
-          </Link>
-        </div>
-        <div className="admin-content">
-          <Switch>
-            <Route path="/admin/contact" render={() => <ContactInfo />} />
-            <Route
-              path="/admin/getinvolved"
-              render={() => <GetInvolved profile={profile} />}
-            />
-            <Route
-              path="/admin/resources"
-              render={() => <Resources resources={resources} />}
-            />
-            <Route
-              path="/admin/events"
-              render={() => <Events events={events} />}
-            />
-            <Route path="/admin" render={() => <Profile />} />
-          </Switch>
-        </div>
-      </div>
-    </div>
+    <ClubPage admin={admin} />
   );
 };
 
 const mapStateToProps = (state) => ({
-  profile: state.profile.profile,
-  events: state.profile.events,
-  resources: state.profile.resources,
+  admin: state.auth.isAuthenticated,
 });
 
 export default connect(mapStateToProps, { loadProfile })(Admin);
