--- src/pages/admin/Admin2.css
@@ -1,9 +1,0 @@
-.container {
-  padding-top: 100px;
-  display: flex;
-  justify-content: center;
-  border: 1px solid black;
-  width: 75vw;
-  margin: 0 auto;
-}
-
