+++ src/pages/admin/AdminOLD.css
@@ -1,0 +1,213 @@
+.clubEdit {
+    height: calc(100vh - 67px);
+    overflow-y: hidden;
+    padding-top: 67px;
+  }
+  
+  .admin-page {
+    display: flex;
+    flex-direction: row;
+  }
+  
+  .admin-sidebar {
+    height: 100vh;
+    display: flex;
+    min-width: 250px;
+    max-width: 250px;
+    padding-top: 20px;
+    padding-bottom: 20px;
+    flex-direction: column;
+    box-shadow: 1px 0.5px 0px rgba(0, 0, 0, 0.07),
+      1px 0.5px 0px rgba(0, 0, 0, 0.15);
+  }
+  
+  .admin-sidebar a {
+    color: #aaa;
+    font-size: 20px;
+    text-decoration: none;
+    padding: 25px 40px;
+    font-weight: bold;
+  }
+  .admin-sidebar a:hover {
+    color: #54a0f1;
+  }
+  
+  .admin-sidebar .selected {
+    color: #54a0f1;
+  }
+  
+  .admin-content {
+    padding: 20px 50px;
+    width: 100%;
+    height: calc(100vh - 64px - 20px - 20px); /* viewport - navbar - padding*/
+    overflow-y: auto;
+  }
+  
+  .admin-page .admin-text {
+    margin-bottom: 30px;
+  }
+  
+  .admin-page .add-resource {
+    padding: 10px 20px;
+  }
+  .admin-page .add-resource .input {
+    display: flex;
+    flex-direction: row;
+    margin: 10px;
+  }
+  .admin-page .input .input-label {
+    width: 100px;
+  }
+  
+  .admin-page .formGroup {
+    display: flex;
+    flex-direction: column;
+    overflow-y: auto;
+    background: #f8f8f8;
+    padding: 30px;
+    padding-top: 30px;
+    margin-bottom: 30px;
+  }
+  
+  .admin-page .formElement {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    width: 100%;
+    line-height: 12px;
+    margin-bottom: 20px;
+  }
+  
+  .admin-page .formElement p {
+    text-align: right;
+    width: 180px;
+    margin-right: 30px;
+    line-height: 20px;
+  }
+  
+  .admin-page .formElement .userInput {
+    font-size: 16px;
+    width: 418px;
+    height: 34px;
+    padding-left: 12px;
+    border: solid 1px #555;
+    border-radius: 5px;
+  }
+  
+  .admin-page .formElement .descriptionInput {
+    width: 500px;
+    height: 200px;
+    padding: 20px;
+    border: solid 1px;
+    border-radius: 5px;
+    resize: none;
+    font-family: 'Helvetica Neue', sans-serif;
+    font-weight: 400;
+    font-style: normal;
+    font-size: 16px;
+    margin-bottom: 4px;
+  }
+  
+  .admin-page .formElement .descriptionInput::placeholder {
+    font-size: 16px;
+  }
+  
+  .admin-page button {
+    margin-bottom: 20px;
+    float: right;
+    background: #54a0f1;
+    border-radius: 5px;
+    padding: 10px 30px;
+    color: white;
+    font-weight: 700;
+    font-size: 16px;
+    border: none;
+    margin-right: 0px;
+  }
+  
+  .admin-page button:hover {
+    background: #6eaff5;
+    cursor: pointer;
+  }
+  
+  .subtitle {
+    margin-top: -20px;
+    font-size: 10px;
+    margin-left: 210px;
+    margin-bottom: 20px;
+    text-decoration: none;
+  }
+  
+  .delete-modal {
+    display: flex;
+    justify-content: center;
+    flex-direction: column;
+    height: 150px;
+    width: 300px;
+    text-align: center;
+    align-items: center;
+  }
+  
+  .admin-page .red {
+    color: #ff2d2d;
+  }
+  .admin-page a {
+    text-decoration: none;
+  }
+  
+  @media screen and (max-width: 850px) {
+    .admin-sidebar {
+      background: white;
+      height: 40px;
+      padding: 10px 0px;
+      position: fixed;
+      top: 67px;
+      left: 0px;
+      max-width: 100vh;
+      width: 100vw;
+      display: table;
+    }
+    .admin-sidebar a {
+      font-size: 12px;
+      padding: 0 10px;
+      display: table-cell;
+      vertical-align: middle;
+      text-align: center;
+    }
+    .admin-content {
+      padding: 0 10px;
+      padding-top: 50px;
+      margin-bottom: 50px;
+    }
+    .admin-page .formGroup {
+      padding: 10px;
+    }
+  
+    .admin-page .formElement {
+      display: flex;
+      flex-direction: column;
+      align-items: unset;
+      width: auto;
+    }
+    .admin-page .formElement p {
+      text-align: left;
+    }
+    .admin-page .formElement .userInput {
+      width: 320px !important;
+      height: 34px;
+    }
+    .subtitle {
+      margin-left: 0px;
+    }
+    .descriptionInput {
+      width: 300px !important;
+    }
+    .react-select__control,
+    .react-select__menu {
+      max-width: 330px !important;
+    }
+    .event {
+      width: 100%;
+    }
+  }
+
